# An assortment of misc small perf changes