They all seem like improvements, but even all together they barely register:
```
       django_template.py             4.4s (2)             4.4s (2)  +0.1%
            pyxl_bench.py             3.9s (2)             3.9s (2)  -0.5%
sqlalchemy_imperative2.py             5.2s (2)             5.2s (2)  -0.4%
        django_migrate.py             1.9s (4)             1.9s (2)  -1.7%
      virtualenv_bench.py             5.2s (2)             5.2s (2)  -0.6%
                  geomean                 3.9s                 3.9s  -0.6%
```